requirements.txt
anaconda==0.0.1.1
argh==0.26.2
asgiref==3.5.0
astroid==2.11.4
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
colorama==0.4.4
coloredlogs==15.0.1
cors==1.0.1
dill==0.3.4
Django==4.0.3
django-baton==2.3.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-ip-logger==1.0.1
django-requestlogs==0.3.0
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
docker-py==1.9.0
filelock==3.6.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
hgicommon==1.2.0
hgijson==1.3.1
humanfriendly==10.0
idna==3.3
inflection==0.5.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
pathtools==0.1.2
Pillow==9.1.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pylint==2.13.8
PySocks==1.7.1
python-dateutil==2.5.3
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.2
tldextract==3.2.1
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
watchdog==0.8.3
websocket-client==1.3.2
whitenoise==6.0.0
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
